Refactor ilm to image_generation
